requirements.txt
python-telegram-bot==21.0.1
openai==1.30.1
python-dotenv==1.0.1
pystray==0.19.5
pillow==10.3.0
schedule==1.2.2